+++ CHANGELOG.md
@@ -13,6 +13,7 @@ For a full diff see [`1.1.0...main`][1.1.0...main].
 * Added `DataProvider\BooleanProvider` ([#326]), by [@localheinz]
 * Added `DataProvider\NullProvider` ([#327]), by [@localheinz]
 * Added `DataProvider\StringProvider` ([#328]), by [@localheinz]
+* Added `DataProvider\IntProvider` ([#335]), by [@localheinz]
 *
 ### Changed
 
@@ -122,6 +123,7 @@ For a full diff see [`0.7.0...0.8.0`][0.7.0...0.8.0].
 [#327]: https://github.com/ergebnis/test-util/pull/327
 [#328]: https://github.com/ergebnis/test-util/pull/328
 [#334]: https://github.com/ergebnis/test-util/pull/334
+[#335]: https://github.com/ergebnis/test-util/pull/335
 
 [@ergebnis]: https://github.com/ergebnis
 [@localheinz]: https://github.com/localheinz